fix: tooltip vista individual Mazo
Corrección del archivo Js para implementar correctamente los tooltip en la vista individual del Mazo.
