+++ snaplabs/src/main/resources/templates/pages/menu_mazos.html
@@ -9,7 +9,7 @@
     <th:block th:fragment="css">
         <link rel="stylesheet" type="text/css" th:href="@{/styles/menu_mazos.css}">
     </th:block>
-    <script th:fragment="js" th:src="@{/js/mazo.js}" type="module" defer></script>
+    <script th:fragment="js" th:src="@{/js/menu_mazos.js}" type="module" defer></script>
 </head>
 <body>
     <main th:fragment="contenido" class="contenido">
